+++ src/components/ui.rs
@@ -42,7 +42,7 @@ impl specs::System<()> for InfoPanelSystem {
             res.window.print_at(Point::new(1, 0), "Health ");
             res.window.print_at(Point::new(1, 1), "Stamina");
             let h = format!("{:30}", format!("{}/{}", health.health, health.max_health));
-            let s = format!("{:30}", format!("{}/{}", health.stamina, health.max_stamina));
+            let s = format!("{:30}", format!("{}/{}", health.stamina as usize, health.max_stamina as usize));
             let mut hfs: FormattedString = (&h).into();
             let mut sfs: FormattedString = (&s).into();
             hfs.bg = Some(ColorValue::Red);
